refactor(cli): split the memory.* /v1 family into cli_v1_routes_e.c
cli_v1_routes.c (2500) and cli_v1_routes_c.c (2499) had both arrived at the
2500-line hard limit that build-integrity and line-check enforce, leaving zero
and one line of headroom. Any further memory work fails CI on file size before
it can be reviewed on merit: #2458 lands them at 2503 and 2506.

Move the memory.* family into a new cli_v1_routes_e.c — the marshallers
(marshal_add_memory_scope through marshal_memory_benchmark) out of
cli_v1_routes.c, and the printers (print_memory_row/search/list/stats and the
pt_print_memory_* entry points) out of cli_v1_routes_c.c. The four existing
files are pure size-overflow buckets with identical preambles; splitting this
one by command family instead means a memory change now touches one file.

The moved code is byte-identical to what it replaces — this commit is a move
and a build-list registration, nothing else. No behaviour changes.

  cli_v1_routes.c    2500 -> 2166
  cli_v1_routes_c.c  2499 -> 2368
  cli_v1_routes_e.c          491

test_cli_v1_delegate.c unity-includes each route file to reach their statics
and links none of the objects, so it gains the matching include.
